% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/academic_year_from_month.R
\name{academic_year_from_month}
\alias{academic_year_from_month}
\title{Compute the year in an academic year from a month}
\usage{
academic_year_from_month(months, ending_year, starting_month = 8L)
}
\arguments{
\item{months}{<int> the months to find the years for}

\item{ending_year}{<int> the calendar year when the academic year ends.}

\item{starting_month}{<int, optional> the first month of the academic year
defaults to August.}
}
\value{
<int> a vector of years suitable for use in e.g. [lubridate::make_date()]
}
\description{
Compute the year in an academic year from a month
}
\examples{
academic_year_from_month(1:12, 2024, 9)
}
