% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assign_race_labels.R
\name{assign_race_labels}
\alias{assign_race_labels}
\title{Convert the race labels from a data source to Racine race labels}
\usage{
assign_race_labels(
  .x,
  mapping,
  ...,
  target_name = "Race/Ethnicity",
  .levels = hiRx::RACE_LABELS_RACINE
)
}
\arguments{
\item{.x}{a data frame with a column of race labels}

\item{mapping}{a data frame that relates the source's labels to Racine's}

\item{...}{values for the `by` argument of an `inner_join` between `.x` and `mapping`}

\item{target_name}{optional, defaults to "Race/Ethnicity".}

\item{.levels}{optional, the new levels for the Race/Ethnicity factor.}
}
\value{
a new data frame with a factor, `Race/Ethnicity`, of Racine races.
}
\description{
Convert the race labels from a data source to Racine race labels
}
