% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BLS_AREA_TYPES}
\alias{BLS_AREA_TYPES}
\title{Metadata about how the US BLS classifies geographic areas}
\format{
a data frame with three columns:
\describe{
\item{prefix}{<chr> the two-letter code for this kind of geograpic area}
\item{digits}{<int> the number of digits this geography needs for its code.}
\item{label}{<chr> a human-readable description of the type of area}
}
}
\source{
\url{https://www.bls.gov/cew/classifications/areas/area-guide.htm}
}
\usage{
BLS_AREA_TYPES
}
\description{
Metadata about how the US BLS classifies geographic areas
}
\keyword{datasets}
