% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_geography.R
\name{bls_geography}
\alias{bls_geography}
\title{Get the BLS string for a specific geographical area}
\usage{
bls_geography(area_type, state_or_region, area = NULL)
}
\arguments{
\item{area_type}{one of several types of area, see notes.}

\item{state_or_region}{the FIPS code of a state or region}

\item{area}{the FIPS code of the sub-region, NULL if unnecessary.}
}
\value{
a 15-digit geocode
}
\description{
Get the BLS string for a specific geographical area
}
