% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_index_table.R
\name{bls_index_table}
\alias{bls_index_table}
\title{Download information about the subtables in a BLS data tree}
\usage{
bls_index_table(table_path)
}
\arguments{
\item{table_path}{the path to the index file, e.g. "pub/time.series/la"}
}
\value{
a data frame with the following fields
    \describe{
        \item{Upload}{the timestamp of when the table was last changed.}
        \item{Bytes}{the size of the table, in bytes}
        \item{Table}{the name of the table}
        \item{Index}{the index of the table, if it contains data}
        \item{Place}{more detail about what the table's data describe}
    }
}
\description{
Download information about the subtables in a BLS data tree
}
