% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_extra_jobs.R
\name{calculate_geo_diff}
\alias{calculate_geo_diff}
\title{Calculate differences across times between two areas for one measure}
\usage{
calculate_geo_diff(.x, .focal_geo, .comp_geo, .measure)
}
\arguments{
\item{.x}{a tibble of LAUS data}

\item{.focal_geo}{the area whose value is on the left side of the subtraction}

\item{.comp_geo}{the area whose value is on the right side}

\item{.measure}{the measure to compare between areas}
}
\value{
a tibble with a new Measure, ``"\{.measure\} Difference"``
}
\description{
Calculate differences across times between two areas for one measure
}
