% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_url.R
\name{census_api_query}
\alias{census_api_query}
\title{Construct the query part of the URL for a call to the US Census API}
\usage{
census_api_query(variables, for_geo, for_items, ...)
}
\arguments{
\item{variables}{a vector of strings with variable codes}

\item{for_geo}{the geographic level of detail that you want}

\item{for_items}{the specific geographic entities that you want}

\item{...}{other key-value pairs to pass to the query}
}
\value{
a string with the query part of a Census API call
}
\description{
Construct the query part of the URL for a call to the US Census API
}
\examples{
census_api_query(paste0("B25003_00", 1:3, "E"),
                 "tract",
                 "*",
                 state = 55L,
                 county = 101L)

census_api_query(paste0("P1_00", c(1, 3, 4), "N"),
                 "block\%20group",
                 "*",
                 state = 55L,
                 county = 101L)
}
