% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_url.R
\name{census_api_url}
\alias{census_api_url}
\title{Create an API call to send to api.census.gov}
\usage{
census_api_url(
  variables,
  for_geo,
  for_items,
  survey_type,
  survey_or_table_code,
  year,
  ...
)
}
\arguments{
\item{variables}{a vector of variables}

\item{for_geo}{a single geographic level, like "tract" or "block"}

\item{for_items}{the specific geographies, or "*"}

\item{survey_type}{e.g. "acs" or "dec"}

\item{survey_or_table_code}{e.g. "acs5" or "pl"}

\item{year}{the year of the survey}

\item{...}{other items to pass to the query}
}
\value{
one URL, as a string
}
\description{
Create an API call to send to api.census.gov
}
\examples{
census_api_url(paste0("B25003_00", 1:3, "E"),
              "tract",
              "*",
              "acs",
              "acs5",
              2020L,
              state = 55L,
              county = 101L)

census_api_url(paste0("P1_00", c(1, 3, 4), "N"),
              "tract",
              "*",
              "dec",
              "pl",
              2020L,
              state = 55L,
              county = 101L)
}
