% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_block_group_coverage.R
\name{census_block_group_coverage}
\alias{census_block_group_coverage}
\title{Find the Percentage of each Covered Census Block Group inside a Geography}
\usage{
census_block_group_coverage(blocks, containing_geo, group_id = "BLKGRPID")
}
\arguments{
\item{blocks}{a simple features table with census block data}

\item{containing_geo}{a simple features table with one geography}

\item{group_id}{tidyselect description of the column with block group ids.}
}
\value{
a tibble with values summarized by block group
}
\description{
Find the Percentage of each Covered Census Block Group inside a Geography
}
