% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_build_acs_variable.R
\name{census_build_acs_variable}
\alias{census_build_acs_variable}
\title{Vectorized Creation of Census ACS Variables}
\usage{
census_build_acs_variable(table_code, race_code, item_number)
}
\arguments{
\item{table_code}{the table code, like "B18101"}

\item{race_code}{either an empty string or one of \[A..I\]}

\item{item_number}{some integer between 0 and 999}
}
\value{
a vector of strings, each of which is a variable name
}
\description{
All detail table variables have the format "B000OOX_000E", where the "B" and
"E" are literal and the "X" is a race code.
}
\examples{
tables <- c("B19013", "B18101", "B18101")
races <- c("", "", "I")
numbers <- 1
census_build_acs_variable(tables, races, numbers)
}
