% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_collapse_races.R
\name{census_collapse_races}
\alias{census_collapse_races}
\title{Consolidate racial categories from the Census's to Higher Expectations'}
\usage{
census_collapse_races(
  .x,
  .census_label_field = "Census Race",
  .value_field = "Population",
  .summary_f = sum
)
}
\arguments{
\item{.x}{a tibble with information grouped by Census race categories.}

\item{.census_label_field}{the name of the field with Census race in it. Defaults to "Census Race"}

\item{.value_field}{the name of the field with numeric data. Defaults to "Population"}

\item{.summary_f}{the function to consolidate values with. Defaults to `sum`.}
}
\value{
a tibble with Higher Expectations' race groups instead of the Census's.
}
\description{
Consolidate racial categories from the Census's to Higher Expectations'
}
