% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_combine_queries.R
\name{census_combine_queries}
\alias{census_combine_queries}
\title{Title}
\usage{
census_combine_queries(
  .l,
  metadata,
  .field_name = "Variable",
  .value_name = "Population",
  .index_name = "Index"
)
}
\arguments{
\item{.l}{a list of JSON-formatted API query responses}

\item{metadata}{a tibble with details about each field in the query}

\item{.field_name}{the column that identifies fields. Defaults to "Variable".}

\item{.value_name}{the column that reports measures. Defaults to "Population".}

\item{.index_name}{the column that indicates a fields' place in a table. Defaults to "Index".}
}
\value{
a tibble of Census data with Higher Expectations factor names.
}
\description{
Title
}
