% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_do_api_query.R
\name{census_do_api_query}
\alias{census_do_api_query}
\title{Fetch JSON-formatted data from the US Census API}
\usage{
census_do_api_query(url, .api_key = NULL)
}
\arguments{
\item{url}{a fully-formed API call in URL form}

\item{.api_key}{an API key}
}
\value{
a list of lists, one row per geo, one entry per variable
}
\description{
Fetch JSON-formatted data from the US Census API
}
