% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_expand_factors.R
\name{census_expand_factors}
\alias{census_expand_factors}
\title{Create a table of census variables from factors and row offsets}
\usage{
census_expand_factors(..., .offset_field = "offset")
}
\arguments{
\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> data frames with factors and offsets}

\item{.offset_field}{the variable that gives row offsets, defaults to "offset"}
}
\value{
a table of census variables with their associated factors and rows
}
\description{
Create a table of census variables from factors and row offsets
}
