% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{census_race_labels}
\alias{census_race_labels}
\title{Codes that flag which race/ethnicity a census table refers to}
\format{
a tibble with 4 columns:
  \describe{
    \item{Race or Ethnic Group}{The name that the census uses for this group}
    \item{Code}{A letter code used to identify census tables and variables.}
    \item{Name}{My preferred name for this socially-defined category of people.}
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
  }
}
\source{
\url{https://www.census.gov/programs-surveys/acs/guidance/which-data-tool/table-ids-explained.html}
}
\usage{
census_race_labels
}
\description{
The census defines several racial and ethnic groups. These arbitrary
categories are somewhat consistent across different census years and
different types of surveys.
}
\keyword{datasets}
