% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_safe_api_query.R
\name{census_safe_api_query}
\alias{census_safe_api_query}
\title{Split a census query into fewer than 50 variables per query}
\usage{
census_safe_api_query(.x, max_per_query = 40, ...)
}
\arguments{
\item{.x}{a tibble with a column called "Variable"}

\item{max_per_query}{defaults to 40}

\item{...}{other arguments for "hiRx::census_api_url"}
}
\value{
a list of JSON-formatted census replies.
}
\description{
The census API complains if you have more than 50 variables per query.
}
