% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_tigershapes_combine.R
\name{census_tigershapes_combine}
\alias{census_tigershapes_combine}
\title{collect shapefiles from TIGER/Line that are from different years.}
\usage{
census_tigershapes_combine(file_names, shapefile_path, ...)
}
\arguments{
\item{file_names}{A named vector or list of shapefiles}

\item{shapefile_path}{The full path to the folder that contains the files}

\item{...}{Arguments that `dplyr::filter` will use on the files' features.}
}
\value{
an `sf::sf` object, a data frame of features with a geometry column
}
\description{
collect shapefiles from TIGER/Line that are from different years.
}
