% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_wrangle_lists_of_raw_json.R
\name{census_wrangle_lists_of_raw_json}
\alias{census_wrangle_lists_of_raw_json}
\title{Convert lists of JSON-formatted census data to tidy tables.}
\usage{
census_wrangle_lists_of_raw_json(
  .x,
  variable_tibble,
  .varname = "Variable",
  .indexname = "Index",
  ...
)
}
\arguments{
\item{.x}{a list of lists of json output from a query to the Census API}

\item{variable_tibble}{a tibble that connects census variables to factor information and row indices}

\item{.varname}{the name of the field that holds Census variable codes}

\item{.indexname}{the name of the field that holds row indices}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> other inputs to \code{\link[hiRx]{census_json_to_tibble}}}
}
\value{
a tibble with factor variables and observed values
}
\description{
Convert lists of JSON-formatted census data to tidy tables.
}
\seealso{
\code{\link[hiRx]{census_expand_factors}}
\code{\link[hiRx]{census_json_to_tibble}}
}
