% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_wrangle_race.R
\name{census_wrangle_race}
\alias{census_wrangle_race}
\title{Condense and relabel racial and ethnic factors}
\usage{
census_wrangle_race(
  .x,
  race_tibble,
  ...,
  .code_field = "Code",
  .name_field = "Name",
  .val_field = "Population"
)
}
\arguments{
\item{.x}{a tibble with Census data that is organized by race}

\item{race_tibble}{a tibble that defines condensed racial/ethnic factors}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> ignored}

\item{.code_field}{defaults to "Code"}

\item{.name_field}{defaults to "Name"}

\item{.val_field}{defaults to "Population"}
}
\value{
a tibble with better factor values
}
\description{
Condense and relabel racial and ethnic factors
}
\seealso{
\code{\link[hiRx]{census_race_labels}}
}
