% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collapse_a_factor.R
\name{collapse_a_factor}
\alias{collapse_a_factor}
\title{Summarize all of the values of a category}
\usage{
collapse_a_factor(
  .x,
  group_to_collapse,
  values_to_summarize,
  collapsed_label,
  .summary_f = sum,
  ...
)
}
\arguments{
\item{.x}{a tibble}

\item{group_to_collapse}{the name of the field whose categories will be collapsed}

\item{values_to_summarize}{the name(s) of the fields with values to summarize}

\item{collapsed_label}{the name of the new, collapsed category}

\item{.summary_f}{the summary function, defaults to `sum`}

\item{...}{further arguments for `.summary_f`, e.g. `na.rm = TRUE`.}
}
\value{
a new summary tibble
}
\description{
Summarize all of the values of a category
}
