% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{COLLEGE_REPORTCARD_API_FIELDS}
\alias{COLLEGE_REPORTCARD_API_FIELDS}
\title{Information about how the US Dept. of Ed reports College Report Card Data}
\format{
a data frame with six columns:
\describe{
\item{dev-category}{<chr>}
\item{developer-friendly name}{<chr>}
\item{API data type}{<chr>}
\item{VARIABLE NAME}{<chr>}
\item{NAME OF DATA ELEMENT}{<chr>}
\item{NOTES}{<chr>}
}
}
\source{
\url{https://collegescorecard.ed.gov/data/api-documentation/}
}
\usage{
COLLEGE_REPORTCARD_API_FIELDS
}
\description{
Information about how the US Dept. of Ed reports College Report Card Data
}
\keyword{datasets}
