% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_table_column_types.R
\name{compare_table_column_types}
\alias{compare_table_column_types}
\title{Align the columns of data frames to see where their names and types synch up}
\usage{
compare_table_column_types(...)
}
\arguments{
\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> named data frames to compare}
}
\value{
a data frame with column names in "Column" and a column of types for each argument
}
\description{
Align the columns of data frames to see where their names and types synch up
}
