% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{CENSUS_RACE_MAPPING}
\alias{CENSUS_RACE_MAPPING}
\alias{CENSUS_RACE_SUFFIXES}
\alias{CHUNK_OPTIONS_FOR_KNITR}
\title{Information for re-coding between Census and Higher Expectations race labels}
\format{
a tibble with three columns
  \describe{
    \item{Census Race}{Labels for each group used by the Census}
    \item{Suffix}{The one-letter code the Census uses to identify tables about each group.}
    \item{Race/Ethnicity}{One of four labels for groups in Racine, WI.}
  }

An object of class \code{character} of length 9.

a list with 8 elements
  \describe{
    \item{`collapse = TRUE`}{Try to put all of the source and output from each block into one code chunk.}
    \item{`warning = FALSE`}{Don't show warning messages in the output.}
    \item{`message = FALSE`}{Don't show informational messages in the output.}
    \item{`comment = "#>"`}{Text appended to every outputted line of code.}
    \item{`cache = TRUE`}{Re-use calculations from chunks that have not changed.}
    \item{`dpi = 96`}{Render the graphics in a web-friendly resolution.}
    \item{`fig.dim = c(width = 7, height = 4)`}{Pleasing defaults for figure sizes.}
    \item{`dev = c("png", "pdf")`}{Cache both `png` and `pdf` outputs, but use pngs in the vignette.}
  }
}
\source{
\url{https://yihui.org/knitr/options/}
}
\usage{
CENSUS_RACE_MAPPING

CENSUS_RACE_SUFFIXES

CHUNK_OPTIONS_FOR_KNITR
}
\description{
In Racine County, WI, only a few racial/ethnic groups occur in large enough
numbers for us to report on their demographics. Even though we lump some
groups into an "All Other Races" category, we still value the cultures and
experiences of people from all heritages.
}
\keyword{datasets}
