% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls_utils.R
\name{count_used_digits}
\alias{count_used_digits}
\title{Count the number of digits that FIPS uses to track a type of area}
\usage{
count_used_digits(.areas)
}
\arguments{
\item{.areas}{The table of area information}
}
\value{
a table of digit counts for each type of area
}
\description{
Count the number of digits that FIPS uses to track a type of area
}
