% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_url.R
\name{decennial_query_url}
\alias{decennial_query_url}
\title{Create an API query URL for a table from a decennial US Census}
\usage{
decennial_query_url(summary_table, ...)
}
\arguments{
\item{summary_table}{the code for a summary table from a decennial census}

\item{...}{<[`dynamic dots`][rlang::dyn-dots]> named arguments to pass to \code{\link[hiRx]{census_api_url}}}
}
\value{
one URL for an API call
}
\description{
Create an API query URL for a table from a decennial US Census
}
\examples{
hiRx::decennial_query_url(variables = paste0("P1_00", c(1, 3, 4), "N"),
                          summary_table = "pl",
                          year = 2020,
                          for_geo = "tract",
                          for_items = "*",
                          state = 55,
                          county = 101)
}
