% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_namespaces.R
\name{detect_namespaces}
\alias{detect_namespaces}
\title{Search a folder for R scripts and list every namespace used in those files}
\usage{
detect_namespaces(.directory)
}
\arguments{
\item{.path}{`<chr>` the directory to search in}
}
\value{
`<chr>` a sorted vector of each namespace
}
\description{
Search a folder for R scripts and list every namespace used in those files
}
