% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emmean_main_effects.R
\name{emmean_main_effects}
\alias{emmean_main_effects}
\title{Create a table of estimates for specific levels of a main effect}
\usage{
emmean_main_effects(.obj, .field)
}
\arguments{
\item{.obj}{`<obj>` a model or [emmeans::emmGrid] object to analyze}

\item{.field}{`<chr>` the name of the main effect to analyze}
}
\value{
`<tbl>` a data frame with
as many rows as there are levels of `.field`
and eight columns
\describe{
  \item{`.field`}{`<chr>` the levels of the main effect}
  \item{emmean}{`<dbl>` the marginal mean for that level of the effect}
  \item{SE}{`<dbl>` the standard error of the marginal mean}
  \item{df}{`<int>` the estimated degrees of freedom for the t test}
  \item{lower.CL}{`<dbl>` the lower bound of the 95\% confidence interval}
  \item{upper.CL}{`<dbl>` the upper bound of the 95\% confidence interval}
  \item{`$t_\{df\}$`}{`<dbl>` the t statistic for this estimate, with df in the name.}
  \item{`$p$-value`}{`<dbl>` the p-value of the t test for this effect}
}
}
\description{
This function is a wrapper for using [emmeans::emmeans()] to examine just
the levels of one main effect from a more complicated model.
It relabels some of the columns for easier use in visualizations.
}
\seealso{
[emmeans::emmeans()]
}
