% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fafsa_read_file.R
\name{fafsa_read_file}
\alias{fafsa_read_file}
\title{Read a file of FAFSA completion data from the US Dept. of Education}
\usage{
fafsa_read_file(path)
}
\arguments{
\item{path}{the full path to the file}
}
\value{
a tibble of helpful information.
}
\description{
Read a file of FAFSA completion data from the US Dept. of Education
}
