% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_synonyms.R
\name{find_synonyms}
\alias{find_synonyms}
\title{Look up the root item that corresponds to each item's synonym}
\usage{
find_synonyms(.x, key_col = "PrimaryKey", syn_col = "Synonym")
}
\arguments{
\item{.x}{a data frame with primary keys and self-referential synonyms}

\item{key_col}{the column that contains the primary keys}

\item{syn_col}{the column that contains self-referential keys}
}
\value{
a vector of primary keys
}
\description{
Look up the root item that corresponds to each item's synonym
}
\examples{
discrete_example <- tibble::tribble(
~ PrimaryKey, ~ Dimension, ~ Order, ~ Label, ~ Synonym,
1L, 1L, 1L, "A", NA,
2L, 1L, 2L, "B", NA,
3L, 1L, 3L, "C", NA,
4L, 1L, NA, "Alpha", 1L,
5L, 1L, NA, "Beta", 2L,
6L, 1L, NA, "Gamma", 3L,
7L, 1L, NA, "\u03B1", 4L
)
synonyms <- find_synonyms(discrete_example)
dplyr::slice(discrete_example, synonyms)
}
