% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_matches.R
\name{fuzzy_matches}
\alias{fuzzy_matches}
\title{Find edit distances between pairwise combinations of two sets of text}
\usage{
fuzzy_matches(.lhs, .rhs, ...)
}
\arguments{
\item{.lhs}{`<chr>` the first set of text values to fuzzy match}

\item{.rhs}{`<chr>` the second set}

\item{...}{
  Arguments passed on to \code{\link[utils:adist]{utils::adist}}
  \describe{
    \item{\code{costs}}{a numeric vector or list with names partially matching
    \samp{insertions}, \samp{deletions} and \samp{substitutions} giving
    the respective costs for computing the Levenshtein distance, or
    \code{NULL} (default) indicating using unit cost for all three
    possible transformations.}
    \item{\code{counts}}{a logical indicating whether to optionally return the
    transformation counts (numbers of insertions, deletions and
    substitutions) as the \code{"counts"} attribute of the return
    value.}
    \item{\code{fixed}}{a logical.  If \code{TRUE} (default), the \code{x}
    elements are used as string literals.  Otherwise, they are taken as
    regular expressions and \code{partial = TRUE} is implied
    (corresponding to the approximate string distance used by
    \code{\link{agrep}} with \code{fixed = FALSE}).}
    \item{\code{partial}}{a logical indicating whether the transformed \code{x}
    elements must exactly match the complete \code{y} elements, or only
    substrings of these.  The latter corresponds to the approximate
    string distance used by \code{\link{agrep}} (by default).}
    \item{\code{ignore.case}}{a logical.  If \code{TRUE}, case is ignored for
    computing the distances.}
    \item{\code{useBytes}}{a logical.  If \code{TRUE} distance computations are
    done byte-by-byte rather than character-by-character.}
  }}
}
\value{
an object of class `tbl_df/tbl/data.frame`
with `length(.lhs)` \eqn{\times} `length(.rhs)` rows and 3 columns:
\describe{
 \item{LHS}{`<chr>` each value of `.lhs`, repeated `len(.rhs)` times}
 \item{RHS}{`<chr>` each value of `.rhs`, repeated `len(.lhs)` times}
 \item{Distance}{`<dbl>` the edit distance between the strings.}
}
}
\description{
The specific kind of edit distance depends upon the extra arguments, which
get passed on to [utils::adist()].
This is really syntactic sugar for converting the matrix that [utils::adist()]
creates into a long tibble.
}
\examples{
fuzzy_matches(c("Foo", "Bar", "Baz"),
              c("Aleph", "Bab", "Jeen", "Dal"))
fuzzy_matches(c("Foo", "Bar", "Baz"),
              c("Aleph", "Bab", "Jeen", "Dal"),
              fixed = FALSE, partial = FALSE, ignore.case = TRUE)
}
