% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GENDER_MAPPING_IPEDS}
\alias{GENDER_MAPPING_IPEDS}
\title{Codes that identify gender in IPEDS demographics tables}
\format{
a tibble with 2 columns:
  \describe{
    \item{Gender}{The IPEDS label for this gender identity.}
    \item{Sex}{A binary label for this set of identities in Racine, WI.}
  }
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
GENDER_MAPPING_IPEDS
}
\description{
Recent files in the US Department of Education's Integrated Postsecondary
Education Database may track a number of different gender identities. This
table oversimplifies them into "Female" and "Non-female" to protect the
identities of Racine people with uncommon gender identities.
}
\keyword{datasets}
