% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_filter_by_interior_intersection.R
\name{geom_filter_by_interior_intersection}
\alias{geom_filter_by_interior_intersection}
\title{Find elements in `atomic_geoms` that overlap with `containing_geom`}
\usage{
geom_filter_by_interior_intersection(atomic_geoms, containing_geom)
}
\arguments{
\item{atomic_geoms}{an `sf::sf` object with one or more features}

\item{containing_geom}{one simple features object}
}
\value{
a tibble with elements from `atomic_geoms`
}
\description{
Find elements in `atomic_geoms` that overlap with `containing_geom`
}
