% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_interiors_intersect.R
\name{geom_interiors_intersect}
\alias{geom_interiors_intersect}
\title{Detect when geometries' interiors intersect.}
\usage{
geom_interiors_intersect(x, y, ...)
}
\arguments{
\item{x}{an object of class `sf::sf`, `sf::sfc` or `sf::sfg`}

\item{y}{an object of class `sf::sf`, `sf::sfc` or `sf::sfg`}

\item{...}{other arguments that might be passed to `sf::st_relate`}
}
\value{
either a matrix of `logical`s or a list of row indices
}
\description{
Detect when geometries' interiors intersect.
}
\seealso{
`sf::st_relate`
}
