% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HERA_SCHOOLS}
\alias{HERA_SCHOOLS}
\title{Identifying information for finding HERA schools in the College Scorecard}
\format{
A tibble with 4 columns:
  \describe{
   \item{UNITID}{Integer code for one specific campus/location}
   \item{OPEID6}{Integer code for one IHD, which may have multiple campuses}
   \item{Name}{Text name for the Institution of Higher Education}
   \item{Campus}{Text name for the campus, optional}
  }
 @source \url{https://collegescorecard.ed.gov/data/}
}
\usage{
HERA_SCHOOLS
}
\description{
The U.S. Department of Education maintains a database of demographic and
performance measures for participating institutions of higher education.
Southeastern Wisconsin has a confederation of such schools, the Higher
Education Regional Alliance.
}
\keyword{datasets}
