% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HERA_SCHOOLS_RACINE}
\alias{HERA_SCHOOLS_RACINE}
\title{Aesthetic and lookup information for three Higher Eds near Racine, WI.}
\format{
A tibble with 4 columns:
  \describe{
   \item{OPEID6}{Integer code for one IHD, which may have multiple campuses}
   \item{Name}{Text name for the Institution of Higher Education}
   \item{Color}{Text representation of the color for plotting each school}
   \item{Shape}{Integer representation of the shape for plotting each school}
  }
 @source \url{https://collegescorecard.ed.gov/data/}
}
\usage{
HERA_SCHOOLS_RACINE
}
\description{
Aesthetic and lookup information for three Higher Eds near Racine, WI.
}
\keyword{datasets}
