% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_ex_cols.R
\name{higher_ex_cols}
\alias{higher_ex_cols}
\title{Look up a color from Higher Expectations' palette}
\usage{
higher_ex_cols(...)
}
\arguments{
\item{...}{Character vectors that name colors. Calling without arguments returns every color.}
}
\value{
An unnamed vector of colors in hex format
}
\description{
Higher Expectations for Racine County has several color palettes that were
designed by the branding team.
}
\note{
The possible values for
}
\examples{
blues <- higher_ex_cols("Bright Blue", "Medium Blue", "Navy")
}
