% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hiRx.R
\docType{package}
\name{hiRx-package}
\alias{hiRx}
\alias{hiRx-package}
\title{hiRx: Utilities for Higher Expectations' data work}
\description{
Right now, there are functions for branding, graphical standards, and data
wrangling
}
\section{Branding}{

Higher Expectations produces reports that use professionally-chosen fonts
and color schemes. These include ...
}

\section{Graphical Standards}{

Higher Expectations' graphs have a consistent look-and-feel. The following
functions facilitate making those graphs.
}

\section{Wrangling Data}{

Higher Expectations makes annual summaries of data that come in reasonably
consistent formats. Data-wrangling functions capture past efforts to deal
with those formats. They are a good place to start when wrangling data from
new sources
}

\seealso{
Useful links:
\itemize{
  \item \url{https://higherx4racine.github.io/hiRx/}
}

}
\author{
\strong{Maintainer}: Ben Taft \email{ben.taft@higherexpectations}

}
