% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_block_group_coverage.R
\name{in_or_out}
\alias{in_or_out}
\title{Pivot blocks to}
\usage{
in_or_out(blocks, containing_geo, group_id)
}
\arguments{
\item{blocks}{the many census blocks}

\item{containing_geo}{the single covering geo}
}
\value{
a wider table
}
\description{
Pivot blocks to
}
