% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_path.R
\name{input_path}
\alias{input_path}
\title{Define a subfolder of the runtime environment's DATA_PATH}
\usage{
input_path(...)
}
\arguments{
\item{...}{path components}
}
\value{
a path to a file or directory
}
\description{
Define a subfolder of the runtime environment's DATA_PATH
}
