% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_percentage_table.R
\name{knit_percentage_table}
\alias{knit_percentage_table}
\title{Print out a table of percentages with nice formatting}
\usage{
knit_percentage_table(.xtable, .groups, .all_field = "All")
}
\arguments{
\item{.xtable}{a data frame produced by [two_way_percentages()]}

\item{.groups}{the column(s) with grouping variables}

\item{.all_field}{the name of the field that holds denominator values}
}
\value{
nothing
}
\description{
Print out a table of percentages with nice formatting
}
\seealso{
[two_way_percentages()]
}
