% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knit_two_ways.R
\name{knit_two_ways}
\alias{knit_two_ways}
\title{Print a nicely-formatted table of percentages}
\usage{
knit_two_ways(.df, .by, .across, .weights)
}
\arguments{
\item{.df}{a table with counts}

\item{.by}{the column that will label each row}

\item{.across}{the column that will be pivoted to columns}

\item{.weights}{the column with in-group counts}
}
\value{
nothing
}
\description{
Print a nicely-formatted table of percentages
}
\seealso{
[two_way_percentages()]

[knit_percentage_table()]
}
