% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_fetch_observations.R
\name{laus_all}
\alias{laus_all}
\title{Download a huge table of LAUS observations}
\usage{
laus_all(.laus_tables)
}
\arguments{
\item{.laus_tables}{a table from calling \code{\link{laus_fetch_table}}.}
}
\value{
A tibble that will have more than a million observations.
   Its columns are:
   \describe{
       \item{series}{a 20-character code, e.g. "LAUCT556600000000003"}
       \item{date}{always the first day of the observation's month}
       \item{value}{a double-precision numeric value}
       \item{footnote_codes}{a one-character code indicating caveats}
   }
   The \code{series} field tracks where the data are from and what they are.
}
\description{
Download a huge table of LAUS observations
}
