% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_extra_jobs.R
\name{laus_extra_jobs}
\alias{laus_extra_jobs}
\title{The number of jobs that would make the two areas' unemployment rates equal.}
\usage{
laus_extra_jobs(.x, .focal_geo, .comp_geo, .result_name = "Additional Jobs")
}
\arguments{
\item{.x}{the tibble with LAUS data}

\item{.focal_geo}{the area whose number of jobs would be changed.}

\item{.comp_geo}{the area whose employment rate is the baseline.}
}
\value{
a tibble with a new category of Measure, "Additional Jobs"
}
\description{
The number of jobs that would make the two areas' unemployment rates equal.
}
