% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_fetch.R
\name{laus_fetch}
\alias{laus_fetch}
\title{Grab data from the BLS API}
\usage{
laus_fetch(
  geographies,
  measures = .laus_measures,
  start_year = 1990L,
  final_year = NULL,
  api_key = NULL
)
}
\arguments{
\item{geographies}{one or more 15-character geocodes}

\item{measures}{one or more members of the measure types enum}

\item{start_year}{the first year of the time series}

\item{final_year}{the last year of the time series, defaults to the current year}

\item{api_key}{a unique id with access to the BLS API}
}
\value{
a data frame
}
\description{
Grab data from the BLS API
}
