% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_fetch_table.R
\name{laus_fetch_table}
\alias{laus_fetch_table}
\title{Get tab-delimited data from download.bls.gov/pub/time.series/la}
\usage{
laus_fetch_table(table_name, column_specification)
}
\arguments{
\item{table_name}{a string that identifies the table to find}

\item{column_specification}{a readr::cols object defining the table's fields.}
}
\value{
a data frame with fields defined by \code{column_specification}
}
\description{
Get tab-delimited data from download.bls.gov/pub/time.series/la
}
