% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LAUS_MEASURE_MAP}
\alias{LAUS_MEASURE_MAP}
\title{Codes that identify the measures being described by a LAUS data series.}
\format{
a tibble with 3 columns:
  \describe{
    \item{Code}{a two-digit character value between 03 and 09}
    \item{Measure}{a name for the measure}
    \item{Description}{a short summary of what the measure quantifies}
  }
}
\source{
\url{https://download.bls.gov/pub/time.series/la/la.measure}
}
\usage{
LAUS_MEASURE_MAP
}
\description{
The U.S. Bureau of Labor Statistics provides monthly reports about Local
Area Unemployment Statistics (LAUS). The measures that it reports are:
}
\keyword{datasets}
