% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_tidy.R
\name{laus_tidy}
\alias{laus_tidy}
\title{Wrangle data downloaded by ``laus_fetch`` into a tidy tibble}
\usage{
laus_tidy(
  .laus_table,
  geo_labels,
  measure_labels = .measure_labels,
  .rate_key = "Rate|Percent"
)
}
\arguments{
\item{.laus_table}{The data from ``laus_fetch``}

\item{geo_labels}{A tibble of human-readable "Geography" values for FIPS codes.}

\item{measure_labels}{A tibble of human-readable names for BLS series IDs.}

\item{.rate_key}{A string to search for to identify percentages.}
}
\value{
a tibble with columns: ``Geography``, ``Measure``, ``Date``, and ``Value``.
}
\description{
Wrangle data downloaded by ``laus_fetch`` into a tidy tibble
}
\examples{
laus_table <- tibble::tribble(
    ~ Series,                ~ Year, ~ Month,   ~ Value,
    "LAUCT556600000000005",    2020,       1,   32436.0,
    "LAUCT556600000000006",    1998,       6,   39679.0,
    "LAUCT556600000000004",    1995,       6,    2760.0,
    "LAUCT556600000000003",    2002,       9,       8.6,
    "LAUCN551010000000005",    1997,      10,   90356.0,
    "LAUCN551010000000006",    2005,       2,   98696.0,
    "LAUCN551010000000004",    2015,      12,    4609.0,
    "LAUCN551010000000003",    2009,      11,      10.0,
    "LAUST550000000000005",    2018,      10, 3040579.0,
    "LAUST550000000000004",    1997,      10,   79483.0,
    "LAUST550000000000003",    2005,       7,       4.7
)
geo_labels <- tibble::tribble(
    ~ State,        ~ Area,     ~ Geography,
       "55", "00000000000",     "Wisconsin",
       "55", "66000000000",   "Racine City",
       "55", "10100000000", "Racine County"
)
print(laus_tidy(laus_table, geo_labels))
}
