% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{make_annotation_labels}
\alias{make_annotation_labels}
\title{Change any percentage values to graphable ones}
\usage{
make_annotation_labels(.x, .rate_key = "Rate")
}
\arguments{
\item{.x}{a tibble}

\item{.rate_key}{a string in the Measure name that indicates a percentage}
}
\value{
a tibble with a new column, ``Label``, that is formatted.
}
\description{
Change any percentage values to graphable ones
}
\examples{
make_annotation_labels(
    tibble::tribble(~Measure, ~Value,
                     "R",      0.5,
                     "K",      1000,
                     "A",      42),
    "R"
)
}
