% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/only_informative_columns.R
\name{only_informative_columns}
\alias{only_informative_columns}
\title{Remove all singleton columns}
\usage{
only_informative_columns(.x)
}
\arguments{
\item{.x}{a tibble or data table with more than one row}
}
\value{
a tibble that may have fewer columns
}
\description{
Remove all singleton columns
}
