% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_levels.R
\name{output_levels}
\alias{output_levels}
\title{Print out quoted, comma-trailed terms, one per line}
\usage{
output_levels(x, ...)
}
\arguments{
\item{x}{the first argument}

\item{...}{additional arguments to be quoted and printed}
}
\value{
nothing, it just prints things
}
\description{
Print out quoted, comma-trailed terms, one per line
}
