% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_location_address.R
\name{parse_location_address}
\alias{parse_location_address}
\title{Extract location address and apartment information from a string with both}
\usage{
parse_location_address(.addresses)
}
\arguments{
\item{.addresses}{<str> - parts of addresses, like "New York, NY 10026"}
}
\value{
a tibble with three columns "City", "State", and "ZIP Code"
}
\description{
Extract location address and apartment information from a string with both
}
