% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_street_address.R
\name{parse_street_address}
\alias{parse_street_address}
\title{Extract street address and apartment information from a string with both}
\usage{
parse_street_address(.addresses)
}
\arguments{
\item{.addresses}{<str> - parts of addresses, like "123 Sesame St. #A"}
}
\value{
a tibble with two columns "Street Address" and "Apartment"
}
\description{
Extract street address and apartment information from a string with both
}
