% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print_array_as_column.R
\name{print_array_as_column}
\alias{print_array_as_column}
\title{Print each element of a vector on a new line, with commas}
\usage{
print_array_as_column(.array, .sep = ",", .quote = "\\"")
}
\arguments{
\item{.array}{<any>, a vector of values, probably string, but could be anything.}

\item{.sep}{<chr>, optional, Delimiter to print before each new line. The default is a comma.}

\item{.quote}{<chr>, optional, Character to wrap each item in. The default is a double quote.}
}
\value{
Nothing, this function is side-effect only.
}
\description{
Print each element of a vector on a new line, with commas
}
\examples{
library(hiRx)
print_array_as_column(c("foo", "bar", "baz"))
}
