% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_mn_fed_cpi.R
\name{pull_mn_fed_cpi}
\alias{pull_mn_fed_cpi}
\title{Scrape the Minneapolis Federal Reserve's Consumer Price Index page.}
\usage{
pull_mn_fed_cpi()
}
\value{
a tibble with 3 columns: year, CPI, and % change.
}
\description{
The Minneapolis fed
[hosts]("https://www.minneapolisfed.org/about-us/monetary-policy/inflation-calculator/consumer-price-index-1913-")
an html table that reports the annual national Consumer Price Index.
}
