% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_LABELS_RACINE}
\alias{RACE_LABELS_RACINE}
\title{Race, at least among Racine Students, usually means Black, Latin@, or White.}
\format{
An ordered list of race labels based on recent ancestry from ...
  \describe{
    \item{Black}{Sub-Saharan Africa}
    \item{Latin@}{Latin America}
    \item{White}{Europe or the Middle East}
    \item{All Other Races}{None of the above}
  }
}
\usage{
RACE_LABELS_RACINE
}
\description{
While folks in town are increasingly likely to identify as multiracial,
there are still a small enough number of them that the only statistically
useful thing to do is to lump everyone else into "All Other Races." This is
obviously bad, but then again, so is all racial categorization? To add to the
perfidy, I'm using the "one drop" system that is so perniciously part of
American racism: you aren't Latin@ if you are Black, and you're not White if
you are Black or Latin@. I am NOT endorsing this system. It reflects the
nature of the community fragmentation that is in Racine today.
}
\keyword{datasets}
