% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_CENSUS}
\alias{RACE_MAPPING_CENSUS}
\title{Information for re-coding between Census and Higher Expectations race labels}
\format{
a tibble with three columns
  \describe{
    \item{SOURCE RACE}{Labels for each group used by the Census}
    \item{Suffix}{The one-letter code the Census uses to identify tables about each group.}
    \item{Race/Ethnicity}{One of four labels for groups in Racine, WI.}
  }
}
\usage{
RACE_MAPPING_CENSUS
}
\description{
In Racine County, WI, only a few racial/ethnic groups occur in large enough
numbers for us to report on their demographics. Even though we lump some
groups into an "All Other Races" category, we still value the cultures and
experiences of people from all heritages.
}
\keyword{datasets}
