% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_DPI_STAFF}
\alias{RACE_MAPPING_DPI_STAFF}
\title{Codes that identify race/ethnicity in WI DPI staff demographics tables}
\format{
a tibble with 2 columns:
  \describe{
    \item{SOURCE RACE}{The one-letter code for this racial label}
    \item{Description}{The WI DPI's short description of this racial group .}
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
  }
}
\source{
\url{https://dpi.wi.gov/wise/data-elements/race}
}
\usage{
RACE_MAPPING_DPI_STAFF
}
\description{
The WI DPI is the Wisconsin Department of Public Instruction. It uses one-
letter codes to report the race/ethnicity of staff members.
}
\keyword{datasets}
