% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_IPEDS}
\alias{RACE_MAPPING_IPEDS}
\title{Codes that identify race/ethnicity in IPEDS demographics tables}
\format{
a tibble with 2 columns:
  \describe{
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
    \item{Population}{The IPEDS label for this racial group.}
  }
}
\source{
\url{https://nces.ed.gov/ipeds/datacenter/DataFiles.aspx}
}
\usage{
RACE_MAPPING_IPEDS
}
\description{
The US Department of Education's Integrated Postsecondary Education
Database only loosely follows the OMB97 definitions of race and ethnicity.
This table maps from its categories to Higher Expectations'.
}
\keyword{datasets}
