% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_PALS}
\alias{RACE_MAPPING_PALS}
\title{Codes that identify race/ethnicity in PALS tables}
\format{
a tibble with 2 columns:
  \describe{
    \item{SOURCE RACE}{An integer code that corresponds to a race.}
    \item{Long Description}{PALS' verbose description of the racial category.}
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
  }
}
\source{
\url{https://palsk8.com/PDF/data-columns/PALS-PreK\%20Data\%20Files\%20Column\%20Headings.pdf}
}
\usage{
RACE_MAPPING_PALS
}
\description{
PALS is the Phonological Awareness and Literacy Screening. It
}
\keyword{datasets}
