% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_TSG}
\alias{RACE_MAPPING_TSG}
\title{Codes that identify race/ethnicity in tables from Teaching Strategies Gold.}
\format{
a tibble with 3 columns:
  \describe{
    \item{Race}{The name that Teaching Strategies uses for this group}
    \item{Ethnicity}{Teaching Strategies tracks Hispanic status with this field.}
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
  }
}
\source{
\url{https://teachingstrategies.force.com/portal/s/article/What-are-the-ethnicity-and-race-selections-for-child-records}
}
\usage{
RACE_MAPPING_TSG
}
\description{
Teaching Strategies LLC, which owns and publishes the GOLD child assessment
suite, has its own set of racial categories that are similar, but not
identical, to the U.S. Census Bureau's categories.
}
\keyword{datasets}
