% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_US_ED_COLLEGE_SCORECARD}
\alias{RACE_MAPPING_US_ED_COLLEGE_SCORECARD}
\title{Codes that identify race/ethnicity in the U.S. Ed. D.'s College Scorecard}
\format{
a tibble with 3 columns:
  \describe{
    \item{SOURCE RACE}{An text abbreviation of the race or ethnicity.}
    \item{Long Description}{A more verbose description of the racial category.}
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
  }
}
\source{
\url{https://collegescorecard.ed.gov/data/}
}
\usage{
RACE_MAPPING_US_ED_COLLEGE_SCORECARD
}
\description{
The United States Department of Education collects a lot of disparate
information into giant tables of "College Scorecards."
}
\keyword{datasets}
