% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_WI_DHS}
\alias{RACE_MAPPING_WI_DHS}
\title{Codes that identify race/ethnicity in COVID data from the WI DHS}
\format{
a tibble with 2 columns:
  \describe{
    \item{SOURCE RACE}{The name that the WI DHS uses for this group}
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
  }
}
\source{
\url{https://data.dhsgis.wi.gov/}
}
\usage{
RACE_MAPPING_WI_DHS
}
\description{
DHS is Wisconsin's Department of Health Services. It has its own set of
racial categories that are similar to, but not identical to, the U.S. Census
Bureau's categories. Crucially, it does not have a "white, non-Hispanic"
racial category. Strictly speaking, that means that Ethnicity and Race
should be two separate variables when dealing with these data.
}
\keyword{datasets}
