% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_MAPPING_WISEDASH}
\alias{RACE_MAPPING_WISEDASH}
\title{Codes that identify race/ethnicity in WISEDash tables}
\format{
a tibble with 2 columns:
  \describe{
    \item{SOURCE RACE}{The name that WISEDash uses for this group}
    \item{Race/Ethnicity}{The broad group for this subgroup in Racine, WI.}
  }
}
\source{
\url{https://dpi.wi.gov/wise/data-elements/race}
}
\usage{
RACE_MAPPING_WISEDASH
}
\description{
WISEDash is the Wisconsin Department of Instruction's data portal. It has
its own set of racial categories that are similar to, but not identical to,
the U.S. Census Bureau's categories.
}
\keyword{datasets}
