% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{racine_race_pop_weights_by_tract}
\alias{racine_race_pop_weights_by_tract}
\title{Populations of Different Races in Census Tracts and Places in Racine, WI}
\format{
a tibble with 6 columns:
    \describe{
        \item{Year}{four-digit integer}
        \item{Municipality}{each county subdivision's NAME field.}
        \item{tract}{the six-digit code of the tract within Racine County}
        \item{Race/Ethnicity}{"Black", "Latin@", "White", or "All Other Races"}
        \item{Municipal Population}{people who live in the tract and Municipality}
        \item{Tract Population}{people in this race who live anywhere in the tract.}
    }
}
\source{
\url{https://api.census.gov}
}
\usage{
racine_race_pop_weights_by_tract
}
\description{
The smallest unit of area that the ACS 5-year survey reports in is the
census tract. No one actually cares about what happens in a census tract
in and of itself, though. County subdivisions, on the other hand, are actual
places like towns, villages, or cities. Those, we care about. This is a
tibble of the populations of three racial/ethnic groups of Racine County, WI
within census tracts and county subdivisions.
}
\keyword{datasets}
