% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_race_mapping.R
\name{read_race_mapping}
\alias{read_race_mapping}
\title{Read a csv file that contains mappings between racial labels}
\usage{
read_race_mapping(.file_name)
}
\arguments{
\item{.file_name}{the full path to the file}
}
\value{
a tibble of character columns
}
\description{
Read a csv file that contains mappings between racial labels
}
