% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scale_aesthetics.R
\name{scale_aesthetics}
\alias{scale_aesthetics}
\alias{scale_color_higher_ex}
\alias{scale_fill_higher_ex}
\title{Scale aesthetics}
\usage{
scale_color_higher_ex(...)

scale_fill_higher_ex(...)
}
\arguments{
\item{...}{further arguments to pass to `grDevices::colorRampPalette`}
}
\value{
the grob being created
}
\description{
Use a Higher Expectations color palette for an aesthetic
}
\examples{
scale_color_higher_ex("logo", n.colors = 2)
scale_color_higher_ex("blues", discrete = FALSE)

scale_fill_higher_ex("tableau", n.colors = 2)
scale_color_higher_ex("blues", discrete = FALSE)

}
