% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_environment_variable.R
\name{set_environment_variable}
\alias{set_environment_variable}
\title{Install a keypair in Your \code{.Renviron} File for Repeated Use}
\usage{
set_environment_variable(key, value, overwrite = FALSE, install = FALSE)
}
\arguments{
\item{key}{The name of the environment variable. TODO: constrain all caps.}

\item{value}{The value of the environment variable.}

\item{overwrite}{If this is set to TRUE, it will overwrite an existing KEY that you already have in your \code{.Renviron} file.}

\item{install}{if TRUE, will install the key in your \code{.Renviron} file for use in future sessions.  Defaults to FALSE.}
}
\description{
This function will add the keypair to your \code{.Renviron}
file so it can be called securely without being stored in your code. After
installation, it can be called any time by typing \code{Sys.getenv("KEY")}
and can be used in package functions by simply typing KEY. If you do not
have an \code{.Renviron} file, the function will create on for you. If you
already have an \code{.Renviron} file, the function will append the key to
your existing file, while making a backup of your original file for disaster
recovery purposes.
}
\examples{

\dontrun{
set_environment_variable("SECRET_KEY" = "111111abc", install = TRUE)
# First time, reload your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("SECRET_KEY")
}

\dontrun{
# If you need to overwrite an existing key:
set_environment_variable("SECRET_KEY", "111111abc", overwrite = TRUE, install = TRUE)
# First time, relead your environment so you can use the key without restarting R.
readRenviron("~/.Renviron")
# You can check it with:
Sys.getenv("SECRET_KEY")
}
}
