% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_block_group_coverage.R
\name{summarize_block_groups}
\alias{summarize_block_groups}
\title{Summarize a collection of block-level geographies}
\usage{
summarize_block_groups(blocks, group_id)
}
\arguments{
\item{blocks}{a simple features table of block-level geographies.}

\item{group_id}{the variable that identifies block groups}
}
\value{
a summarized table
}
\description{
Summarize a collection of block-level geographies
}
