% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/to_camel_case.R
\name{to_camel_case}
\alias{to_camel_case}
\title{Convert  strings to CamelCase strings}
\usage{
to_camel_case(.strings)
}
\arguments{
\item{.strings}{*&lt;chr&gt;* strings with word breaks instead of spaces.}
}
\value{
*&lt;chr&gt;* the same strings without spaces and capitalized letters at the starts of words.
}
\description{
Convert  strings to CamelCase strings
}
