% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_way_percentages.R
\name{two_way_percentages}
\alias{two_way_percentages}
\title{Compute the percentage of some total found in two margins}
\usage{
two_way_percentages(.df, .by, .across, .weights)
}
\arguments{
\item{.df}{a data frame with counts and categories}

\item{.by}{the category that will stay as a factor column}

\item{.across}{the category that will become columns}

\item{.weights}{the column that counts individuals}
}
\value{
a table with rows determined by .by and columns by .across
}
\description{
Compute the percentage of some total found in two margins
}
