% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bls.R
\name{unlump_series}
\alias{unlump_series}
\title{Break up a BLS series id into its spatial and contextual components}
\usage{
unlump_series(.x, .series)
}
\arguments{
\item{.x}{a data frame with at least one column of BLS series codes.}

\item{.series}{a character vector of 20-digit BLS series codes}
}
\value{
a tibble with 5 columns
}
\description{
Break up a BLS series id into its spatial and contextual components
}
\examples{
library(magrittr)
racine_laus \%>\%
    dplyr::bind_rows() \%>\%
    dplyr::select(Series) \%>\%
    dplyr::distinct() \%>\%
    unlump_series(Series)
}
