% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{US_ED_COLLEGE_SCORECARD_SPEC}
\alias{US_ED_COLLEGE_SCORECARD_SPEC}
\title{Default columns for importing data from the US College Scorecard}
\format{
A list of `readr::col_*` specifications.
}
\source{
\url{https://collegescorecard.ed.gov/data/}
}
\usage{
US_ED_COLLEGE_SCORECARD_SPEC
}
\description{
Default columns for importing data from the US College Scorecard
}
\keyword{datasets}
