% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_valid_intersections.R
\name{valid_intersections}
\alias{valid_intersections}
\title{Only report intersections with valid geometries}
\usage{
valid_intersections(geos_of_interest, reference_geo)
}
\arguments{
\item{geos_of_interest}{a collection of `sf::sf`-like objects}

\item{reference_geo}{a single `sf::sf` object}
}
\value{
a collection of new, valid `sf::sf` objects
}
\description{
Only report intersections with valid geometries
}
\seealso{
\describe{
    \item{`sf::st_intersection`}{the function for computing the intersections}
    \item{`sf::st_make_valid`}{the function that repairs the new geometries}
}
}
