% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{annotation_table}
\alias{annotation_table}
\title{Create a data from for putting labels on a graph.}
\usage{
annotation_table(.x, .col, .focal_value)
}
\arguments{
\item{.x}{the data frame to summarize}

\item{.col}{the column to find indices in.}

\item{.focal_value}{a value to focus on (e.g. the year 2020)}
}
\value{
a data frame with fields ``Name``, ``Measure``, ``Value``, and ``Label``
}
\description{
Create a data from for putting labels on a graph.
}
\examples{
racine_laus \%>\%
    purrr::pluck("wisconsin") \%>\%
    hiRx::unlump_series(Series) \%>\%
    dplyr::filter(`Measure Code` == "06", Month == "January") \%>\%
    identify_indices(Year, 2020)
}
