% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{annotation_table}
\alias{annotation_table}
\title{Create a data frame for putting labels on a graph.}
\usage{
annotation_table(.x, .col, .focal_value, .group_col)
}
\arguments{
\item{.x}{<obj> the data frame to summarize}

\item{.col}{<str> the name of the column to find indices in.}

\item{.focal_value}{<num> a value to focus on (e.g. the year 2020)}

\item{.group_col}{<str> the name of the variable to find indices within}
}
\value{
a data frame with fields ``Name``, ``Measure``, ``Value``, and ``Label``
}
\description{
Create a data frame for putting labels on a graph.
}
\examples{
racine_laus |>
    purrr::pluck("wisconsin") |>
    dplyr::mutate(Parsed = list(laus_parse_series(.data$Series))) |>
    tidyr::unnest("Parsed") |>
    dplyr::select("measure_code", "Year", "Month", "Value") |>
    dplyr::filter(.data$Month == "January") |>
    annotation_table("Year", 2020, "measure_code")
}
