% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_expand_factors.R
\name{census_expand_factors}
\alias{census_expand_factors}
\title{Create a table of census variables from factors and row offsets}
\usage{
census_expand_factors(
  .x,
  .level_field = "Level",
  .offset_field = "Offset",
  ...
)
}
\arguments{
\item{.x}{the table of factors and row offsets}

\item{.level_field}{the variable that names factor levels}

\item{.offset_field}{the variable that gives row offsets}

\item{...}{<[`dynamic-dots`][rlang::dyn-dots]> other factors without offsets}
}
\value{
a table of census variables with their associated factors and rows
}
\description{
Create a table of census variables from factors and row offsets
}
