% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_json_to_tibble.R
\name{census_json_to_tibble}
\alias{census_json_to_tibble}
\title{Wrangle Census Table JSON}
\usage{
census_json_to_tibble(.l)
}
\arguments{
\item{.l}{a list of lists from fetching Census data in JSON format}
}
\value{
a tibble in which rows are geometries and columns are variables.
}
\description{
Wrangle Census Table JSON
}
