% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/census_api_query.R
\name{decennial_query_url}
\alias{decennial_query_url}
\title{Title}
\usage{
decennial_query_url(year, summary_table, variables, for_geo, for_items, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
