% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/higher_ex_pal.R
\name{higher_ex_pal}
\alias{higher_ex_pal}
\title{Look up a Higher Expectations color palette by name}
\usage{
higher_ex_pal(palette = "main", reverse = FALSE, n.colors = NULL, ...)
}
\arguments{
\item{palette}{the name of the palette, defaults to "main"}

\item{reverse}{should the palette's order be reversed, defaults to FALSE}

\item{n.colors}{choose this to limit the number of colors in the palette, defaults to NULL for all of the palette's colors}

\item{...}{further arguments to pass to `grDevices::colorRampPalette`}
}
\value{

}
\description{
Look up a Higher Expectations color palette by name
}
\examples{
higher_ex_pal("blues")
}
