% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{identify_indices}
\alias{identify_indices}
\title{Identify the indices of rows that will be labelled in a graph}
\usage{
identify_indices(.x, .col, .focal_value, .peak_col = "Value")
}
\arguments{
\item{.x}{the data frame to summarize}

\item{.col}{the column to find indices in.}

\item{.focal_value}{a value to focus on (e.g. the year 2020)}

\item{.peak_col}{a column to search for its max. defaults to "Value"}
}
\value{
a tibble with ``Name``, ``Index``, ``X Offset``, and ``Y Offset``
}
\description{
Identify the indices of rows that will be labelled in a graph
}
\examples{
racine_laus \%>\%
    purrr::pluck("wisconsin") \%>\%
    hiRx::unlump_series(Series) \%>\%
    dplyr::filter(`Measure Code` == "06", Month == "January") \%>\%
    identify_indices(Year, 2020)
}
