% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/laus_parse_series.R
\name{laus_parse_series}
\alias{laus_parse_series}
\title{Separate a LAUS series code into its components}
\usage{
laus_parse_series(.x, .col = "series_id")
}
\arguments{
\item{.x}{a data frame or character vector with series codes}

\item{.col}{optional. If \code{.x} is a data frame, this is the name of the
column that holds the series codes. It defaults to "series".}
}
\value{
a data frame with the following fields
    \describe{
        \item{seasonal}{'S' for seasonally-transformed data, 'U' otherwise}
        \item{area_prefix}{a 2-character code describing a level of geographic organization}
        \item{srd_code}{an integer holding a 2-digit FIPS code for a state, census region, or census division}
        \item{area_fips}{an 11-character FIPS code for an intrastate area}
        \item{measure_code}{an integer code for what the series measures.}
    }
}
\description{
Separate a LAUS series code into its components
}
\examples{
tibble::tribble(
    ~ series,
    "LAUST290000000000009",
    "LAUCS233569500000005",
    "LAUMT542658000000003",
    "LAUCN170630000000005",
    "LAUCA392120000000004",
    "LAUCN210230000000003",
    "LAUCN190410000000006",
    "LAUCT134319200000005",
    "LAUMC554802000000004",
    "LAURD840000000000006"
)
}
