% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_table.R
\name{make_annotation_labels}
\alias{make_annotation_labels}
\title{Create a column of formatted labels from one of disparate numeric values.}
\usage{
make_annotation_labels(.x, .rate_key = "Rate")
}
\arguments{
\item{.x}{a tibble with fields ``Measure`` and ``Value``.}

\item{.rate_key}{a string in the Measure name that indicates a percentage}
}
\value{
a tibble with a new column, ``Label``, that is formatted.
}
\description{
This function looks in a character column for a regex that indicates whether
a row is a percentage or a whole number. It then prettily formats the
contents of a numeric column as either a percentage or a comma-separated int.
}
\examples{
make_annotation_labels(
    tibble::tribble(~Measure, ~Value,
                     "R",      0.5,
                     "K",      1000,
                     "A",      42),
    "R"
)
}
