% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RACE_LABELS_RACINE}
\alias{RACE_LABELS_RACINE}
\title{Race, at least among Racine Students, usually means Black, Latin@,
Multiracial, or White.}
\format{
An ordered list of race labels based on recent ancestry from ...
  \describe{
    \item{Black}{Sub-Saharan Africa}
    \item{Hispanic or Latin@}{former colonies of Spain, including most of Latin America}
    \item{Multiracial}{More than one of the other three categories.}
    \item{White and Other Races}{Europe, Asia, Oceania, and non-Hispanic indigenous peoples of the Western Hemisphere}
  }
}
\usage{
RACE_LABELS_RACINE
}
\description{
The vast majority of students in Racine Unified identify with one of those
four racial/ethnic identities. What about folks who identify with locally
less frequent racial categories, such as Asian, Native American, or Pacific
Islander? In the area of Racine Unified School District, the distribution of
socioeconomic situations that those folks find themselves in is most similar
to that of folks who identify as white. As of 2025, therefore, we will use
the following 4 categories when reporting results with racial identities.
}
\keyword{datasets}
