% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_college_scorecard.R
\name{read_college_scorecard}
\alias{read_college_scorecard}
\title{Import data downloaded from the National College Scorecard}
\usage{
read_college_scorecard(.path)
}
\arguments{
\item{.path}{the full path to the file}
}
\value{
a data frame / tibble with a lot of juicy detail about enrollment
}
\description{
The U.S. Department of Education maintains a database of information about
institutions of higher education. You can download big Excel files from
\url{https://collegescorecard.ed.gov/data/}.
}
