% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/week_of_month.R
\name{week_of_month}
\alias{week_of_month}
\title{Compute week numbers of dates for making calendars}
\usage{
week_of_month(.days, .wdays = NULL)
}
\arguments{
\item{.days}{`<int|Date>` days of the month or dates}

\item{.wdays}{`<int?>` weekdays (Sunday == 1), required when `.days` is an integer}
}
\value{
`<int>` row numbers (starting at 1) for plotting a calendar
}
\description{
Compute week numbers of dates for making calendars
}
\examples{
week_of_month(lubridate::ymd(c("2026-01-01", "2026-02-01", "2026-03-01")))
}
